feat: Setup modules for DB connection
I added a sample endpoint to work with. This does not have any JWT
authentication yet, that's for another PR. This has two types of
queries:

    * `authenticate_query/3` for queries that need to switch roles; and
    * `query/2` for queries that only need the `hp_anon` role.

These are convenience functions that will set the necessary things
needed for each query transaction, so that I don't have to do it over
and over again.